requirements.txt
flask>=3.0.2
ultralytics>=8.2.0
opencv-python-headless>=4.9.0.80
pillow>=10.4.0 
numpy>=1.26.4
reportlab>=4.1.0
python-dotenv>=1.0.1